refactor(BA-6608): move ExistenceCheck/SearchScope to models.scopes
These search-scope abstractions are parameterized by ORM columns and produce
QueryConditions, so they belong at the models layer rather than repositories.
Add models/scopes.py, drop them from repositories/base (definition and
re-export), and import them directly from models.scopes everywhere.
IntegrityErrorCheck stays in repositories/base/types.py (it depends on
errors.repository).
