+++ src/ai/backend/manager/repositories/app_config_fragment/types.py
@@ -12,7 +12,7 @@ from ai.backend.manager.errors.app_config import AppConfigDefinitionNotFound
 from ai.backend.manager.models.app_config_definition.row import AppConfigDefinitionRow
 from ai.backend.manager.models.app_config_fragment.row import AppConfigFragmentRow
 from ai.backend.manager.models.clauses import QueryCondition
-from ai.backend.manager.repositories.base import ExistenceCheck, SearchScope
+from ai.backend.manager.models.scopes import ExistenceCheck, SearchScope
 
 __all__ = ("ConfigNameSearchScope",)
