+++ src/ai/backend/manager/repositories/audit_log/db_source/db_source.py
@@ -12,10 +12,10 @@ from ai.backend.common.resilience.policies.retry import BackoffStrategy, RetryAr
 from ai.backend.common.resilience.resilience import Resilience
 from ai.backend.manager.data.audit_log.types import AuditLogData, AuditLogListResult
 from ai.backend.manager.models.audit_log import AuditLogRow
+from ai.backend.manager.models.scopes import SearchScope
 from ai.backend.manager.repositories.base import (
     BatchQuerier,
     Creator,
-    SearchScope,
     execute_batch_querier,
     execute_creator,
 )
