+++ src/ai/backend/manager/repositories/audit_log/types.py
@@ -11,7 +11,7 @@ import sqlalchemy as sa
 from ai.backend.common.data.permission.types import RBACElementType
 from ai.backend.manager.models.audit_log import AuditLogRow
 from ai.backend.manager.models.clauses import QueryCondition
-from ai.backend.manager.repositories.base import ExistenceCheck, SearchScope
+from ai.backend.manager.models.scopes import ExistenceCheck, SearchScope
 
 __all__ = (
     "EntityAuditLogSearchScope",
