+++ src/ai/backend/manager/repositories/auth/types.py
@@ -11,8 +11,8 @@ import sqlalchemy as sa
 from ai.backend.common.exception import UserNotFound
 from ai.backend.manager.models.clauses import QueryCondition
 from ai.backend.manager.models.login_session.row import LoginHistoryRow, LoginSessionRow
+from ai.backend.manager.models.scopes import ExistenceCheck, SearchScope
 from ai.backend.manager.models.user import UserRow
-from ai.backend.manager.repositories.base.types import ExistenceCheck, SearchScope
 
 
 @dataclass(frozen=True)
