+++ src/ai/backend/manager/repositories/base/querier.py
@@ -12,9 +12,9 @@ import sqlalchemy as sa
 from ai.backend.manager.errors.repository import UnsupportedCompositePrimaryKeyError
 from ai.backend.manager.models.base import Base
 from ai.backend.manager.models.clauses import QueryCondition, QueryOrder
+from ai.backend.manager.models.scopes import ExistenceCheck, SearchScope
 
 from .pagination import PageInfoResult, QueryPagination
-from .types import ExistenceCheck, SearchScope
 
 if TYPE_CHECKING:
     from sqlalchemy.engine import Row
