+++ src/ai/backend/manager/repositories/base/types.py
@@ -2,12 +2,9 @@
 
 from __future__ import annotations
 
-from abc import ABC, abstractmethod
-from collections.abc import Callable, Sequence
+from collections.abc import Callable
 from dataclasses import dataclass
 from typing import TYPE_CHECKING, Any, TypeVar
-
-import sqlalchemy as sa
 
 from ai.backend.common.exception import BackendAIError
 from ai.backend.manager.models.clauses import QueryCondition
@@ -16,50 +13,6 @@ if TYPE_CHECKING:
     from sqlalchemy.engine import Row
 
     from ai.backend.manager.errors.repository import RepositoryIntegrityError
-
-
-T = TypeVar("T")
-
-
-@dataclass(frozen=True)
-class ExistenceCheck[T]:
-    """Defines an existence check for scope validation.
-
-    Used to validate that required entities exist before executing a query.
-    Multiple checks are combined into a single query for efficiency.
-    """
-
-    column: sa.orm.attributes.InstrumentedAttribute[T]
-    """The column to check (e.g., ScalingGroupRow.name)."""
-
-    value: T
-    """The value to check for existence."""
-
-    error: BackendAIError
-    """The error to raise if the entity doesn't exist."""
-
-
-class SearchScope(ABC):
-    """Abstract base class for search scope.
-
-    Scope defines required parameters for entity-based search queries.
-    It converts to a QueryCondition that can be added to BatchQuerier conditions.
-    Optionally defines existence checks for validation.
-    """
-
-    @abstractmethod
-    def to_condition(self) -> QueryCondition:
-        """Convert scope to a query condition."""
-        raise NotImplementedError
-
-    @property
-    @abstractmethod
-    def existence_checks(self) -> Sequence[ExistenceCheck[Any]]:
-        """Return existence checks for scope validation.
-
-        All checks are validated in a single query before the main query executes.
-        """
-        raise NotImplementedError
 
 
 # Factory function that creates a cursor condition from a decoded cursor value (str or UUID)
