+++ src/ai/backend/manager/repositories/deployment/types/endpoint.py
@@ -28,7 +28,7 @@ from ai.backend.manager.errors.resource import ProjectNotFound
 from ai.backend.manager.models.clauses import QueryCondition
 from ai.backend.manager.models.endpoint.row import EndpointRow
 from ai.backend.manager.models.group.row import GroupRow
-from ai.backend.manager.repositories.base.types import ExistenceCheck, SearchScope
+from ai.backend.manager.models.scopes import ExistenceCheck, SearchScope
 
 
 @dataclass
