+++ src/ai/backend/manager/repositories/domain/types.py
@@ -11,7 +11,7 @@ from ai.backend.manager.data.domain.types import DomainData
 from ai.backend.manager.errors.resource import ScalingGroupNotFound
 from ai.backend.manager.models.clauses import QueryCondition
 from ai.backend.manager.models.scaling_group import ScalingGroupForDomainRow, ScalingGroupRow
-from ai.backend.manager.repositories.base import ExistenceCheck, SearchScope
+from ai.backend.manager.models.scopes import ExistenceCheck, SearchScope
 
 __all__ = (
     "DomainSearchScope",
