+++ src/ai/backend/manager/repositories/keypair/types.py
@@ -16,8 +16,8 @@ from ai.backend.manager.errors.user import UserNotFound
 from ai.backend.manager.models.clauses import QueryCondition
 from ai.backend.manager.models.keypair.row import KeyPairRow
 from ai.backend.manager.models.resource_policy.row import KeyPairResourcePolicyRow
+from ai.backend.manager.models.scopes import ExistenceCheck, SearchScope
 from ai.backend.manager.models.user.row import UserRow
-from ai.backend.manager.repositories.base import ExistenceCheck, SearchScope
 
 __all__ = (
     "UserKeypairSearchScope",
