+++ src/ai/backend/manager/repositories/role_invitation/types.py
@@ -11,7 +11,7 @@ import sqlalchemy as sa
 from ai.backend.manager.data.role_invitation.types import RoleInvitationData
 from ai.backend.manager.models.clauses import QueryCondition
 from ai.backend.manager.models.role_invitation.row import RoleInvitationRow
-from ai.backend.manager.repositories.base import ExistenceCheck, SearchScope
+from ai.backend.manager.models.scopes import ExistenceCheck, SearchScope
 
 
 @dataclass
