+++ src/ai/backend/manager/repositories/session/types.py
@@ -11,8 +11,8 @@ import sqlalchemy as sa
 from ai.backend.manager.errors.resource import ProjectNotFound
 from ai.backend.manager.models.clauses import QueryCondition
 from ai.backend.manager.models.group.row import GroupRow
+from ai.backend.manager.models.scopes import ExistenceCheck, SearchScope
 from ai.backend.manager.models.session.row import SessionRow
-from ai.backend.manager.repositories.base import ExistenceCheck, SearchScope
 
 __all__ = ("ProjectSessionSearchScope",)
