+++ src/ai/backend/manager/repositories/vfolder/types.py
@@ -14,9 +14,9 @@ from ai.backend.manager.errors.resource import ProjectNotFound
 from ai.backend.manager.errors.user import UserNotFound
 from ai.backend.manager.models.clauses import QueryCondition
 from ai.backend.manager.models.group import GroupRow
+from ai.backend.manager.models.scopes import ExistenceCheck, SearchScope
 from ai.backend.manager.models.user.row import UserRow
 from ai.backend.manager.models.vfolder import VFolderPermissionRow, VFolderRow
-from ai.backend.manager.repositories.base import ExistenceCheck, SearchScope
 
 __all__ = (
     "BulkVFolderPurgeResult",
