+++ src/ai/backend/manager/services/auth/actions/search_login_sessions.py
@@ -4,8 +4,8 @@ from typing import override
 from ai.backend.manager.actions.action import SearchActionResult
 from ai.backend.manager.actions.types import ActionOperationType
 from ai.backend.manager.data.auth.login_session_types import LoginSessionData
+from ai.backend.manager.models.scopes import SearchScope
 from ai.backend.manager.repositories.base.querier import BatchQuerier
-from ai.backend.manager.repositories.base.types import SearchScope
 from ai.backend.manager.services.auth.actions.base import AuthAction
