+++ tests/unit/manager/repositories/ops/test_provider.py
@@ -19,17 +19,16 @@ from sqlalchemy.orm import Mapped, mapped_column
 from ai.backend.manager.errors.repository import EmptySearchScopeError
 from ai.backend.manager.models.base import Base
 from ai.backend.manager.models.clauses import QueryCondition
+from ai.backend.manager.models.scopes import ExistenceCheck, SearchScope
 from ai.backend.manager.models.utils import ExtendedAsyncSAEngine
 from ai.backend.manager.repositories.base import (
     BatchQuerier,
     Creator,
     CreatorSpec,
     DependentCreatorSpec,
-    ExistenceCheck,
     NoPagination,
     Purger,
     Querier,
-    SearchScope,
     Updater,
     UpdaterSpec,
 )
